- reversed App.js to function component - added new Client and join room call to BottomTabNavigator as the top-level component of the game to pass room to screens - updated HomeScreen and RoleSelectionScreen
