+++ client/App.js
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import PropTypes from 'prop-types';
 import { Platform, StatusBar, StyleSheet, View } from 'react-native';
 import { SplashScreen } from 'expo';
 import * as Font from 'expo-font';
@@ -10,118 +9,59 @@ import { createStackNavigator } from '@react-navigation/stack';
 import BottomTabNavigator from './navigation/BottomTabNavigator';
 import useLinking from './navigation/useLinking';
 import { NightTheme } from "./constants/Colors";
-import { Client } from "colyseus.js";
 
 const Stack = createStackNavigator();
 
-class App extends React.Component {
-  static propTypes = {
-    skipLoadingScreen: PropTypes.bool,
-  };
+export default function App(props) {
+  const [isLoadingComplete, setLoadingComplete] = React.useState(false);
+  const [initialNavigationState, setInitialNavigationState] = React.useState();
 
-  static defaultProps = {
-    skipLoadingScreen: true,
-  };
+  const containerRef = React.useRef();
+  const { getInitialState } = useLinking(containerRef);
 
-  constructor(props) {
-    super(props);
+  // Load any resources or data that we need prior to rendering the app
+  React.useEffect(() => {
+    async function loadResourcesAndDataAsync() {
+      try {
+        SplashScreen.preventAutoHide();
 
-    this.container = null;
+        // Load our initial navigation state
+        setInitialNavigationState(await getInitialState());
 
-    this.state = {
-      isLoading: false,
-      navigation: null,
-      client: null,
-      room: null,
-    };
-  }
+        // Load fonts
+        await Font.loadAsync({
+          ...Ionicons.font,
+          'space-mono': require('./assets/fonts/SpaceMono-Regular.ttf'),
 
-  async componentDidMount() {
-    await this.start();
-  }
-
-  componentWillUnmount() {
-    this.stop();
-  }
-
-  start = async () => {
-    let navigation, room;
-    try {
-      const host = window.document.location.host.replace(/:.*/, '');
-      const port = process.env.NODE_ENV !== 'production' ? '2567' : window.location.port;
-      const url = window.location.protocol.replace('http', 'ws') + '//' + host + (port ? ':' + port : '');
-
-      this.client = new Client(url);
-      room = await this.client.joinOrCreate('my_room');
-
-      navigation = await useLinking(this.container);
-
-      SplashScreen.preventAutoHide();
-
-      // Load fonts
-      await Font.loadAsync({
-        ...Ionicons.font,
-        'space-mono': require('./assets/fonts/SpaceMono-Regular.ttf'),
-
-      });
-      await Font.loadAsync({
-        'werewolf': require('./assets/fonts/Werewolf.ttf'),
-      });
-
-      // this.room.onStateChange(state => this.loadPlayers(state));
-    } catch (e) {
-      // We might want to provide this error information to an error reporting service
-      console.warn('Fucked by:', e);
-    } finally {
-      this.setState({ isLoading: true, navigation, room });
-      SplashScreen.hide();
+        });
+        await Font.loadAsync({
+          'werewolf': require('./assets/fonts/Werewolf.ttf'),
+        });
+      } catch (e) {
+        // We might want to provide this error information to an error reporting service
+        console.warn('Fucked by:', e);
+      } finally {
+        setLoadingComplete(true);
+        SplashScreen.hide();
+      }
     }
-  };
 
-  stop = () => {
-    // Colyseus
-    if (this.room) {
-      this.room.leave();
-    }
-  };
-  //
-  // loadPlayers = async (state) => {
-  //   const serverPlayers = state.players;
-  //
-  //   let players = [];
-  //   for (let id in serverPlayers) {
-  //     let player = serverPlayers[id];
-  //     players.push({ id: id, ...player });
-  //   }
-  //
-  //   this.setState({ players });
-  // };
+    loadResourcesAndDataAsync();
+  }, []);
 
-  registerContainerRef = (element) => {
-    this.container = element;
-  };
-
-  render() {
-    const { isLoading, navigation, room } = this.state;
-    // const { client, room } = this.props;
-
-    console.debug("Client is: ", this.client);
-    console.debug("Room is: ", room);
-
-    if (isLoading && !this.props.skipLoadingScreen) {
-      return null;
-    } else {
-      return (
-        <View style={styles.container}>
-          {Platform.OS === 'ios' && <StatusBar barStyle="default" />}
-          <NavigationContainer ref={this.registerContainerRef} initialState={navigation}>
-            <Stack.Navigator>
-              <Stack.Screen name="Root" component={BottomTabNavigator} />
-            </Stack.Navigator>
-          </NavigationContainer>
-        </View>
-      );
-    }
+  if (!isLoadingComplete && !props.skipLoadingScreen) {
+    return null;
+  } else {
+    return (
+      <View style={styles.container}>
+        {Platform.OS === 'ios' && <StatusBar barStyle="default" />}
+        <NavigationContainer ref={containerRef} initialState={initialNavigationState}>
+          <Stack.Navigator>
+            <Stack.Screen name="Root" component={BottomTabNavigator} />
+          </Stack.Navigator>
+        </NavigationContainer>
+      </View>
+    );
   }
 }
 
@@ -131,5 +71,3 @@ const styles = StyleSheet.create({
     backgroundColor: NightTheme.darkBlue,
   },
 });
-
-export default App;
